menambahkan edit dan tambah di ruangan,user, dan fixedschedules

+++ src/api/rooms.js
@@ -2,7 +2,8 @@ import api from './axios'
 
 export const rooms = {
   all: (params = {}) => api.get('/rooms', { params }),
-  create: (data) => api.post('/rooms', data),
-  update: (id, data) => api.put(`/rooms/${id}`, data),
-  destroy: (id) => api.delete(`/rooms/${id}`),
+  create: (data) => api.post('/admin/rooms', data),
+  get: (id) => api.get(`/rooms/${id}`),
+  update: (id, data) => api.put(`/admin/rooms/${id}`, data),
+  destroy: (id) => api.delete(`/admin/rooms/${id}`),
 }
